# clients/docs/src/main/openapi/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "ActivityInfo API 3.0",
    "description": "ActivityInfo 3.0 AP",
    "version": "1.0.0"
  },
  "host": "www.activityinfo.org",
  "schemes": [
    "https"
  ],
  "basePath": "/resources",
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/form/{formId}/schema": {
      "get": {
        "operationId": "getFormSchema",
        "summary": "Get a Form's Schema",
        "tags": [
          "forms"
        ],
        "description": "Retrieves the schema that describes this form's fields.\nA form in ActivityInfo is a set of records that have a common set of fields, or schema. \nA form can be a user-defined form created by a user, or it can be an application-defined like the set of\n countries, which has the id `_countries`.\n",
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the form",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The form's schema",
            "schema": {
              "$ref": "#/definitions/FormSchema"
            }
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form"
          },
          "404": {
            "description": "The form does not exist."
          }
        }
      },
      "put": {
        "operationId": "updateFormSchema",
        "summary": "Updates a Form's Schema",
        "tags": [
          "forms"
        ],
        "description": "Updates the form\n",
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the Form",
            "required": true,
            "type": "string"
          },
          {
            "name": "updatedSchema",
            "in": "body",
            "description": "Updates the schema describing this form's fields",
            "schema": {
              "$ref": "#/definitions/FormSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The form's schema was updated"
          },
          "400": {
            "description": "The update was invalid. Not all changes to a schema are permitted. For example, a field's type cannot be changed."
          },
          "401": {
            "description": "The authenticated user does not have permission to update this form"
          },
          "404": {
            "description": "The specified form does not exist"
          }
        }
      }
    },
    "/form/{formId}/form.xls": {
      "get": {
        "operationId": "getFormSchemaXls",
        "x-gwt-client": false,
        "summary": "Get a Form's Schema as XLSForm",
        "tags": [
          "forms"
        ],
        "description": "Retrieves a Form's Schema, formatted as an <a href=\"http://xlsform.org\">XLSForm</a>. XLSForm are Excel\nWorkbooks that contain a tab called \"survey\" with the list of fields, and a tab called \"choices\" that contain\nthe options for single and multiple select fields.\n",
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the Form",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/vnd.ms-excel"
        ],
        "responses": {
          "200": {
            "description": "The form's schema, converted to an XLSForm"
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form"
          },
          "404": {
            "description": "The form does not exist."
          }
        }
      }
    },
    "/form/{formId}/tree": {
      "get": {
        "operationId": "getSchemaTree",
        "x-gwt-client": false,
        "x-doc": false,
        "summary": "Get a Form's Schema Tree",
        "description": "Retrieves this form's schema, together with the schemas of all related forms.\nThe fields defined in a form's schema may reference other forms, which in turn may refer to yet others.\nAs a convenience, this request returns the schemas of all forms reachable from this form.\n",
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the form",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The form's class",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FormSchema"
              }
            }
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form"
          },
          "404": {
            "description": "The form does not exist."
          }
        }
      }
    },
    "/form/{formId}/records": {
      "get": {
        "operationId": "getRecords",
        "summary": "Get All Records",
        "tags": [
          "records"
        ],
        "description": "Gets all records belonging to a Form.",
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the Form containing the Record",
            "required": true,
            "type": "string"
          },
          {
            "name": "parentId",
            "in": "query",
            "description": "Limits the records to those which are sub-records of this parent record id.",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested records",
            "schema": {
              "$ref": "#/definitions/FormRecordSet"
            }
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form"
          },
          "404": {
            "description": "No form with `formId` exists"
          }
        }
      },
      "post": {
        "operationId": "createRecord",
        "summary": "Create a New Record",
        "tags": [
          "records"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the Form",
            "required": true,
            "type": "string"
          },
          {
            "name": "newRecord",
            "in": "body",
            "description": "The record to create",
            "schema": {
              "$ref": "#/definitions/NewFormRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record was updated"
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form"
          },
          "404": {
            "description": "No form with `formId` exists."
          }
        }
      }
    },
    "/form/{formId}/record/{recordId}": {
      "get": {
        "operationId": "getRecord",
        "summary": "Get a Single Form Record",
        "tags": [
          "records"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the Form",
            "required": true,
            "type": "string"
          },
          {
            "name": "recordId",
            "in": "path",
            "description": "Id of the record",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The requested record",
            "schema": {
              "$ref": "#/definitions/FormRecord"
            }
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form or record"
          },
          "404": {
            "description": "No record with `formId` exists"
          }
        }
      },
      "put": {
        "operationId": "updateRecord",
        "summary": "Update a Form Record",
        "tags": [
          "records"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the Form",
            "required": true,
            "type": "string"
          },
          {
            "name": "recordId",
            "in": "path",
            "description": "Id of the record",
            "type": "string",
            "required": true
          },
          {
            "name": "update",
            "in": "body",
            "description": "The record to update",
            "schema": {
              "$ref": "#/definitions/FormRecordUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record was updated"
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form or record"
          },
          "404": {
            "description": "No record with `formId` exists"
          }
        }
      }
    },
    "/form/{formId}/record/{recordId}/history": {
      "get": {
        "operationId": "getRecordHistory",
        "x-doc-section": "records",
        "summary": "Get a Record's History",
        "tags": [
          "records"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "description": "Id of the Form",
            "required": true,
            "type": "string"
          },
          {
            "name": "recordId",
            "in": "path",
            "description": "Id of the record",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The requested record",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FormHistoryEntry"
              }
            }
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form or record"
          },
          "404": {
            "description": "No record with `formId` exists"
          }
        }
      }
    },
    "/query/columns": {
      "post": {
        "operationId": "queryTableColumns",
        "summary": "Query Table Columns",
        "tags": [
          "query"
        ],
        "deescription": "Flattens records from one or more forms into a two-dimensional table, returning the results as a list of columns. Though forms can contain rich, nested and linked data, for the purposes of analysis a simple two-dimensional table is often required.",
        "parameters": [
          {
            "name": "query",
            "in": "body",
            "description": "The shape of the table to retrieve",
            "schema": {
              "$ref": "#/definitions/TableQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The table as a list of column",
            "schema": {
              "$ref": "#/definitions/ColumnSet"
            }
          },
          "401": {
            "description": "The authenticated user does not have permission to view this form"
          },
          "404": {
            "description": "No form with `formId` exists."
          }
        }
      }
    }
  },
  "definitions": {
    "FormSchema": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Unique identifier of this class."
        },
        "databaseId": {
          "type": "string",
          "description": "The id of the database to which this form belongs."
        },
        "parentFormId": {
          "type": "string",
          "description": "The id of the parent form, if this is a subform"
        },
        "label": {
          "type": "string",
          "description": "Short, human-readable name of the class"
        },
        "code": {
          "type": "string",
          "description": "short identifier that can be used to refer to the field in formulas"
        },
        "description": {
          "type": "string",
          "description": "Description of this class"
        },
        "elements": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Field"
          }
        }
      },
      "required": [
        "id",
        "databaseId",
        "label",
        "elements"
      ]
    },
    "Field": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Unique, immutable identifier of this field."
        },
        "code": {
          "type": "string",
          "description": "Short identifier"
        },
        "label": {
          "type": "string",
          "description": "Name of the field"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "description": "The type of value in this field",
          "type": "string",
          "enum": [
            "text",
            "quantity"
          ]
        },
        "typeParameters": {
          "description": "Parameters of the field type.",
          "type": "object"
        },
        "relevanceCondition": {
          "description": "A boolean-valued expression that determines whether the field is shown",
          "type": "string"
        },
        "required": {
          "description": "True if a value for this field is required",
          "type": "boolean"
        },
        "visible": {
          "description": "True if the field should be displayed during data entry",
          "type": "boolean"
        }
      }
    },
    "FormRecord": {
      "type": "object",
      "properties": {
        "recordId": {
          "type": "string",
          "description": "This record's id."
        },
        "formId": {
          "type": "string",
          "description": "The id of the form to which this record belongs"
        },
        "parentRecordId": {
          "type": "string",
          "description": "The id of the parent FormRecord, if this record belongs to a subform"
        },
        "keyId": {
          "type": "string",
          "description": "The key id, if this record belongs to a subform"
        },
        "fields": {
          "type": "object",
          "description": "Dictionary of field values of this FormRecord"
        }
      }
    },
    "FormRecordSet": {
      "type": "object",
      "properties": {
        "formId": {
          "type": "string",
          "description": "The id of the form to which this record belongs"
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FormRecord"
          }
        }
      }
    },
    "FormRecordUpdate": {
      "type": "object",
      "properties": {
        "deleted": {
          "type": "boolean",
          "description": "True if the record should be deleted"
        },
        "fieldValues": {
          "type": "object",
          "description": "Dictionary of the fields to update"
        }
      }
    },
    "NewFormRecord": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "client-generated id"
        },
        "parentRecordId": {
          "type": "string",
          "description": "id of the parent FormRecord, if this is record is a member of a subform"
        },
        "keyId": {
          "type": "string",
          "description": "key id, if this is record is a member of a subform"
        },
        "fieldValues": {
          "type": "object",
          "description": "Dictionary of the fields intial values"
        }
      }
    },
    "TableQuery": {
      "type": "object",
      "properties": {
        "rowSources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TableRowSource"
          }
        },
        "columns": {
          "$ref": "#/definitions/ColumnModel"
        },
        "filter": {
          "type": "string",
          "description": "An boolean-valued expression to select which resources to include in the table"
        },
        "required": "-rowSources -columns"
      }
    },
    "TableRowSource": {
      "type": "object",
      "properties": {
        "formId": {
          "type": "string",
          "description": "The id of a form to include as a source of rows for this table"
        }
      }
    },
    "ColumnModel": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "An arbitrary id to give this column in the resulting `ColumnSet`"
        },
        "expression": {
          "type": "string",
          "description": "The query expression"
        }
      }
    },
    "ColumnSet": {
      "type": "object",
      "properties": {
        "numRows": {
          "type": "integer",
          "description": "The number of rows in each column"
        },
        "columns": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ColumnView"
          }
        }
      }
    },
    "ColumnView": {
      "type": "object",
      "properties": {
        "numRows": {
          "type": "integer",
          "description": "The number of rows in this column"
        },
        "values": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "FormHistoryEntry": {
      "type": "object",
      "properties": {
        "formId": {
          "type": "string",
          "description": "id of the form"
        },
        "recordId": {
          "type": "string",
          "description": "id of the record"
        },
        "time": {
          "type": "integer",
          "description": "the time, in seconds since 1970-01-01, that the change was made"
        },
        "subFieldId": {
          "type": "string",
          "description": "for sub records, the subForm field to which this sub record belongs"
        },
        "subFieldLabel": {
          "type": "string",
          "description": "for sub records, the label of the subForm field to which this sub record belongs"
        },
        "subRecordKey": {
          "type": "string",
          "description": "for keyed sub forms, such as monthly, weekly, or daily subForms, this is a human readable label describing the key, for example '2016-06'"
        },
        "changeType": {
          "type": "string",
          "enum": [
            "created",
            "updated",
            "deleted"
          ]
        },
        "userName": {
          "type": "string",
          "description": "the name of the user who made the change"
        },
        "userEmail": {
          "type": "string",
          "description": "the email address of the user who made the change"
        },
        "values": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FormValueChange"
          }
        }
      }
    },
...
  }
}